add to cart data fetch changes
