+++ index.html
@@ -178,5 +178,6 @@
       <p class="footer-copyright">Copyright © 2024. All rights reserved.</p>
     </footer>
     <script src="/script.js"></script>
+    <script src="/src/js/addToCartFunc.js" type="module"></script>
   </body>
 </html>
