+++ src/js/cart.js
@@ -1,84 +1,22 @@
-if (!localStorage.getItem("cartList")) { 
-    localStorage.setItem("cartList", JSON.stringify([]))
-}
-let cartCount = document.querySelector(".count") 
-let productContainer = document.getElementById("posts") // change element when add to official code
-let globalData = null
 
-// -------------------------------------------------------------------------------------------------- 
-
-async function getData() {
-    try {
-        const response = await fetch("https://fakestoreapi.com/products")
-        const data = await response.json()
-        globalData = data
-        console.log(globalData);
-        showItems(data)
-     }
-    catch (error) {
-       console.error("Fetch error:" , error)
-     }
-}
-
-getData()
-// -------------------------------------------------------------------------------------------------- 
-// -------------------------------------------------------------------------------------------------- 
-
-const showItems = (passData) => {  
-    productContainer.innerHTML = passData.map(items => (` 
-        <article class="post" id=${items.id} >
-        <img src="${items.image}" alt="product" height="100px" width = "100px" >
-        <h2>${items.title}</h2>
-        <p>${items.description}</p>
-        <p>${items.category}</p>
-        <p>${items.price}</p>
-        <button onclick = getById(${items.id})>Add to cart</button>
-        </article>
-        `)). join(" ")
-    }
-// -------------------------------------------------------------------------------------------------- 
-// --------------------------------------------------------------------------------------------------     
+let getCartItems = JSON.parse(localStorage.getItem("cartList"))
+let listContainer = document.querySelector(".ul")
 
 
-
-let cartItemCountLS = () => {
-    const getLSitems = JSON.parse(localStorage.getItem("cartList"))
-    console.log( getLSitems.length);
-
- }
-
-const getById = (id) => { 
-    if (!globalData) {  // checks if the globaldata variable is populated or not
-        console.error("data not loaded yet") 
-        return null
-    }
-
-    const items = globalData.find((item) => item.id === id) //if globaldata is populated then find the item with matching id
-    if (!items) {                                           // compares the id from the clicked item with and looks for the same one trought globaldata
-        console.warn(`item with id:${id} is not found `);
-        return null
-    }
-    const getCartLS = JSON.parse(localStorage.getItem("cartList"))
-          getCartLS.push(items)
-          localStorage.setItem("cartList", JSON.stringify(getCartLS))
-    
-          for (let i = 0; i < getCartLS.length; i++) {
-            cartItemCountLS()
-            cartCount.innerHTML = i + 1;
-         }
-    console.log("found item", items);
-    console.log("item is added to cart");
-    return items
-    }
-
-
-
-       
-
-    
-
-document.addEventListener("DOMContentLoaded", () => { 
-
-})
-
+    listContainer.innerHTML = getCartItems.map(items => (
+        `<li class="CartItem" id=${items.id}>
+                            <img class="itemImage" src="${items.image}" width="100px" height="100px" alt="">
+                            <div class="itemTitleWrapper">
+                                <span class="itemTitle">${items.title}</span>
+                                <p class="itemDescription">${items.description}</p>
+                            </div>
+                            <div class="itemCountWrapper">
+                                <button class="itemCountBtn"> - </button>
+                                <span class="itemCountCounter"></span>
+                                <button class="itemCountBtn"> + </button>
+                            </div>
+                            <h3 class="itemPrice">$ ${items.price} </h3>
+                            <img id="removeItemBtn" src="/Webshop icons/icons8-trash-128.png"/>
+                        </li>`
+    )).join(" ")
